services/catalogue/requirements.txt
Flask>=3.0.0
gunicorn>=20.1.0
requests>=2.26.0
flask_cors>=3.0.10
psycopg2-binary
python-dotenv>=1.0.0
